2.Main Activity cleaned from other classes
