+++ app/src/main/java/com/gb/stopwatch/TimestampMillisecondsFormatter.kt
@@ -1,0 +1,25 @@
+package com.gb.stopwatch
+
+class TimestampMillisecondsFormatter {
+
+    fun format(timestamp: Long): String {
+        val millisecondsFormatted = (timestamp % 1000).pad(3)
+        val seconds = timestamp / 1000
+        val secondsFormatted = (seconds % 60).pad(2)
+        val minutes = seconds / 60
+        val minutesFormatted = (minutes % 60).pad(2)
+        val hours = minutes / 60
+        return if (hours > 0) {
+            val hoursFormatted = (minutes / 60).pad(2)
+            "$hoursFormatted:$minutesFormatted:$secondsFormatted"
+        } else {
+            "$minutesFormatted:$secondsFormatted:$millisecondsFormatted"
+        }
+    }
+
+    private fun Long.pad(desiredLength: Int) = this.toString().padStart(desiredLength, '0')
+
+    companion object {
+        const val DEFAULT_TIME = "00:00:000"
+    }
+}
