+++ app/src/main/java/com/gb/stopwatch/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
